updated col sizes.
For some reason the sizing is off, it should be col-lg-4 and then col-md-6 but it doesn't lay out correctly when I do it that way. I'm assuming it's something to do with margins or padding. It works if I have no margins, so that must be the only case it works that way, which makes sense.